requirements.txt
flask
jwt
requests
lxml
flask_sqlalchemy
flask_marshmallow
short_url
sqlalchemy_utils
redis
flask_testing
mock